Implemented camt052 and refactored code to reuse existing components
