+++ lib/camt_parser.rb
@@ -7,13 +7,15 @@ require "bigdecimal"
 require_relative "camt_parser/misc"
 require_relative "camt_parser/version"
 require_relative "camt_parser/errors"
-require_relative "camt_parser/053/account_balance"
-require_relative "camt_parser/053/group_header"
+require_relative "camt_parser/general/account_balance"
+require_relative "camt_parser/general/creditor"
+require_relative "camt_parser/general/debitor"
+require_relative "camt_parser/general/entry"
+require_relative "camt_parser/general/account"
+require_relative "camt_parser/general/group_header"
+require_relative "camt_parser/052/report"
+require_relative "camt_parser/052/base"
 require_relative "camt_parser/053/statement"
-require_relative "camt_parser/053/creditor"
-require_relative "camt_parser/053/debitor"
-require_relative "camt_parser/053/entry"
-require_relative "camt_parser/053/account"
 require_relative "camt_parser/053/base"
 require_relative "camt_parser/file"
 require_relative "camt_parser/string"
