+++ spec/fixtures/052/valid_example.xml
@@ -1,0 +1,132 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Document xmlns="urn:iso:std:iso:20022:tech:xsd:camt.052.001.02" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="urn:iso:std:iso:20022:tech:xsd:camt.052.001.02 camt.052.001.02.xsd">
+
+<BkToCstmrAcctRpt>
+  <GrpHdr>
+    <MsgId>052D2013-12-27T11:02:03.0N130000005</MsgId>
+    <CreDtTm>2013-12-27T11:01:46.0+01:00</CreDtTm>
+    <MsgPgntn>
+      <PgNb>1</PgNb>
+      <LastPgInd>true</LastPgInd>
+    </MsgPgntn>
+  </GrpHdr>
+  <Rpt>
+    <Id>0352C5220131227110203</Id>
+    <ElctrncSeqNb>130000005</ElctrncSeqNb>
+    <CreDtTm>2013-12-27T11:01:46.0+01:00</CreDtTm>
+    <Acct>
+      <Id>
+        <IBAN>DE58740618130100033626</IBAN>
+      </Id>
+      <Ccy>EUR</Ccy>
+      <Ownr>
+        <Nm>Testkonto Nummer 2</Nm>
+      </Ownr>
+      <Svcr>
+        <FinInstnId>
+          <BIC>GENODEF1PFK</BIC>
+          <Nm>VR-Bank Rottal-Inn eG</Nm>
+          <Othr>
+            <Id>DE 129267947</Id>
+            <Issr>UmsStId</Issr>
+          </Othr>
+        </FinInstnId>
+      </Svcr>
+    </Acct>
+    <Bal>
+      <Tp>
+        <CdOrPrtry>
+          <Cd>PRCD</Cd>
+        </CdOrPrtry>
+      </Tp>
+      <Amt Ccy="EUR">8.50</Amt>
+      <CdtDbtInd>CRDT</CdtDbtInd>
+      <Dt>
+        <Dt>2013-12-27</Dt>
+      </Dt>
+    </Bal>
+    <Bal>
+      <Tp>
+        <CdOrPrtry>
+          <Cd>CLBD</Cd>
+        </CdOrPrtry>
+      </Tp>
+      <Amt Ccy="EUR">13.50</Amt>
+      <CdtDbtInd>CRDT</CdtDbtInd>
+      <Dt>
+        <Dt>2013-12-27</Dt>
+      </Dt>
+    </Bal>
+    <Ntry>
+      <Amt Ccy="EUR">2.00</Amt>
+      <CdtDbtInd>CRDT</CdtDbtInd>
+      <Sts>BOOK</Sts>
+      <BookgDt>
+        <Dt>2013-12-27</Dt>
+      </BookgDt>
+      <ValDt>
+        <Dt>2013-12-27</Dt>
+      </ValDt>
+      <AcctSvcrRef>2013122710583450000</AcctSvcrRef>
+      <BkTxCd/>
+      <NtryDtls>
+        <TxDtls>
+          <BkTxCd>
+            <Prtry>
+              <Cd>NMSC+051</Cd>
+              <Issr>ZKA</Issr>
+            </Prtry>
+          </BkTxCd>
+          <RltdPties>
+            <Dbtr>
+              <Nm>Testkonto Nummer 1</Nm>
+            </Dbtr>
+            <DbtrAcct>
+              <Id>
+                <Othr>
+                  <Id>  740618130000033626</Id>
+                  <SchmeNm>
+                    <Cd>BBAN</Cd>
+                  </SchmeNm>
+                </Othr>
+              </Id>
+            </DbtrAcct>
+          </RltdPties>
+          <RmtInf>
+            <Ustrd>TEST BERWEISUNG MITTELS BLZ UND KONTONUMMER - DTA</Ustrd>
+          </RmtInf>
+        </TxDtls>
+      </NtryDtls>
+    </Ntry>
+    <Ntry>
+      <Amt Ccy="EUR">3.00</Amt>
+      <CdtDbtInd>CRDT</CdtDbtInd>
+      <Sts>BOOK</Sts>
+      <BookgDt>
+        <Dt>2013-12-27</Dt>
+      </BookgDt>
+      <ValDt>
+        <Dt>2013-12-27</Dt>
+      </ValDt>
+      <AcctSvcrRef>2013122710583670000</AcctSvcrRef>
+      <BkTxCd/>
+      <NtryDtls>
+        <TxDtls>
+          <Refs>
+            <EndToEndId>NOTPROVIDED</EndToEndId>
+          </Refs>
+          <BkTxCd>
+            <Prtry>
+              <Cd>NMSC+051</Cd>
+              <Issr>ZKA</Issr>
+            </Prtry>
+          </BkTxCd>
+          <RmtInf>
+            <Ustrd>Testüberweisung mit BIC und IBAN SEPA IBAN: DE14740618130000033626 BIC: GENODEF1PFK</Ustrd>
+          </RmtInf>
+        </TxDtls>
+      </NtryDtls>
+    </Ntry>
+  </Rpt>
+</BkToCstmrAcctRpt>
+</Document>
