+++ spec/fixtures/052/valid_namespace.xml
@@ -1,0 +1,4 @@
+<?xml version='1.0' encoding='UTF-8' ?>
+<Document xmlns="urn:iso:std:iso:20022:tech:xsd:camt.052.001.02">
+
+</Document>
