+++ spec/fixtures/general/invalid_namespace.xml
