+++ spec/lib/camt_parser/053/statement_spec.rb
@@ -10,12 +10,12 @@ describe CamtParser::Format053::Statement do
   specify { expect(ex_stmt.generation_date).to be_kind_of(Time) }
   specify { expect(ex_stmt.from_date_time).to be_nil }
   specify { expect(ex_stmt.to_date_time).to be_nil }
-  specify { expect(ex_stmt.account).to be_kind_of(CamtParser::Format053::Account) }
+  specify { expect(ex_stmt.account).to be_kind_of(CamtParser::Account) }
   specify { expect(ex_stmt.entries).to be_kind_of(Array) }
   specify { expect(ex_stmt.electronic_sequence_number).to eq('130000005') }
 
-  specify { expect(ex_stmt.opening_balance).to be_kind_of(CamtParser::Format053::AccountBalance) }
-  specify { expect(ex_stmt.closing_balance).to be_kind_of(CamtParser::Format053::AccountBalance) }
+  specify { expect(ex_stmt.opening_balance).to be_kind_of(CamtParser::AccountBalance) }
+  specify { expect(ex_stmt.closing_balance).to be_kind_of(CamtParser::AccountBalance) }
 
   specify { expect(ex_stmt.identification).to eq("0352C5320131227220503") }
 end
