+++ spec/lib/camt_parser/general/account_balance_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe CamtParser::Format053::AccountBalance do
+describe CamtParser::AccountBalance do
   let(:camt)       { CamtParser::File.parse('spec/fixtures/053/valid_example.xml') }
   let(:statements) { camt.statements }
   let(:ex_stmt)    { camt.statements[0] }
