+++ spec/lib/camt_parser/general/account_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe CamtParser::Format053::Account do
+describe CamtParser::Account do
   let(:camt)       { CamtParser::File.parse('spec/fixtures/053/valid_example.xml') }
   let(:statements) { camt.statements }
   let(:ex_stmt)    { camt.statements[0] }
